+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Master
   * Added `queueSize` option to `initPool` to set the request overfow queue size 
+  * Added option to supply `cdnURL` to build script (#133)
 
 # 2.0.14